+++ tests/index_generator/test_np_index_generator.py
@@ -4,6 +4,7 @@ from pathlib import Path
 
 import numpy as np
 import pytest
+import torch
 
 from connectlib.index_generator.np_index_generator import NpIndexGenerator
 
@@ -11,7 +12,14 @@ from connectlib.index_generator.np_index_generator import NpIndexGenerator
 @pytest.mark.parametrize("batch_size", [0, 3, 5, 8])
 def test_np_index_generator_drop_last(batch_size):
     # Check that if the last batch is dropped, all batches always have the same size, i.e. the batch size.
-    nig = NpIndexGenerator(n_samples=10, batch_size=batch_size, shuffle=True, drop_last=True, seed=42)
+    nig = NpIndexGenerator(
+        n_samples=10,
+        batch_size=batch_size,
+        shuffle=True,
+        drop_last=True,
+        seed=42,
+        num_updates=20,
+    )
     for _ in range(9):
         assert len(nig.__next__()) == batch_size, "All batches do not have the same size when drop_last is set to True."
 
@@ -27,6 +35,7 @@ def test_np_index_generator_keep_last(n_samples, batch_size):
         shuffle=True,
         drop_last=False,
         seed=42,
+        num_updates=2000,
     )
 
     n_batch_per_epoch = nig._n_batch_per_epoch
@@ -61,6 +70,7 @@ def test_np_index_generator_100_epoch(n_samples, batch_size, drop_last):
         shuffle=True,
         drop_last=drop_last,
         seed=42,
+        num_updates=2000,
     )
     for _ in range(100 * nig._n_batch_per_epoch):
         assert len(nig.__next__()) > 0
@@ -76,6 +86,7 @@ def test_np_index_generator_n_samples_negative():
             shuffle=True,
             drop_last=True,
             seed=42,
+            num_updates=40,
         )
 
 
@@ -89,6 +100,7 @@ def test_np_index_generator_batch_size_negative():
             shuffle=True,
             drop_last=True,
             seed=42,
+            num_updates=40,
         )
 
 
@@ -101,14 +113,14 @@ def test_np_index_generator_batch_size_negative():
 )
 def test_np_index_generator_batch_shuffle(seed, results):
     # Check that the shuffling is properly seeded.
-    nig = NpIndexGenerator(n_samples=3, batch_size=3, shuffle=True, drop_last=True, seed=seed)
+    nig = NpIndexGenerator(n_samples=3, batch_size=3, shuffle=True, drop_last=True, seed=seed, num_updates=40)
     for result in results:
         assert np.array_equal(result, nig.__next__())
 
 
 def test_np_index_generator_not_batch_shuffle():
     # Check that we can have not shuffled batches.
-    nig = NpIndexGenerator(n_samples=3, batch_size=3, shuffle=False, drop_last=True, seed=42)
+    nig = NpIndexGenerator(n_samples=3, batch_size=3, shuffle=False, drop_last=True, seed=42, num_updates=40)
     res = np.arange(3)
 
     for _ in range(10):
@@ -127,6 +139,7 @@ def test_np_index_generator_check_epoch_consitency(n_samples, batch_size, drop_l
         shuffle=True,
         drop_last=drop_last,
         seed=42,
+        num_updates=40,
     )
 
     n_epoch = nig.n_epoch_generated
@@ -155,6 +168,7 @@ def test_np_index_generator_count_batch_per_epoch(n_samples, batch_size, drop_la
         shuffle=True,
         drop_last=drop_last,
         seed=42,
+        num_updates=50,
     )
     n_epoch = 0
 
@@ -170,7 +184,7 @@ def test_np_index_generator_statefulness(session_dir):
     # Check that after saving and loading, the batch generator still has the same outputs.
     indexer_path = Path(tempfile.mkdtemp(dir=session_dir)) / "my_indexer"
 
-    nig = NpIndexGenerator(n_samples=5, batch_size=3, shuffle=True, drop_last=True, seed=42)
+    nig = NpIndexGenerator(n_samples=5, batch_size=3, shuffle=True, drop_last=True, seed=42, num_updates=50)
 
     # Artificial 10 first batch
     for _ in range(10):
@@ -189,3 +203,66 @@ def test_np_index_generator_statefulness(session_dir):
     # Check that both of them returns the same results
     for _ in range(20):
         assert np.array_equal(nig.__next__(), loaded_nig.__next__())
+
+
+def test_np_index_generator_num_updates():
+    """Check that the iterator stops after num_updates iterations"""
+    nig = NpIndexGenerator(n_samples=3, batch_size=3, shuffle=True, drop_last=True, seed=10, num_updates=10)
+    for _ in range(10):
+        next(nig)
+    with pytest.raises(StopIteration):
+        next(nig)
+
+
+def test_np_index_generator_counter():
+    """Check that the counter is incremented at each call"""
+    nig = NpIndexGenerator(n_samples=3, batch_size=3, shuffle=True, drop_last=True, seed=10, num_updates=10)
+    assert nig.counter == 0
+    next(nig)
+    assert nig.counter == 1
+
+
+def test_np_index_generator_reset():
+    """Test the reset function"""
+    num_updates = 10
+    nig = NpIndexGenerator(n_samples=3, batch_size=3, shuffle=True, drop_last=True, seed=10, num_updates=num_updates)
+    for _ in range(num_updates):
+        next(nig)
+
+    nig.reset_counter()
+    assert nig.counter == 0
+    for _ in range(num_updates):
+        next(nig)
+    with pytest.raises(StopIteration):
+        next(nig)
+
+
+@pytest.mark.parametrize("num_workers", list(range(6)))
+def test_np_index_generator_torch_dataloader(num_workers):
+    """Test that the index generator works with a torch dataloader"""
+    n_samples = 100
+    num_updates = 10
+    nig = NpIndexGenerator(
+        n_samples=n_samples, batch_size=3, shuffle=True, drop_last=True, seed=10, num_updates=num_updates
+    )
+    nig_copy = NpIndexGenerator(
+        n_samples=n_samples, batch_size=3, shuffle=True, drop_last=True, seed=10, num_updates=num_updates
+    )
+
+    class DummyDataset(torch.utils.data.Dataset):
+        def __init__(self):
+            self.data = np.asarray(range(n_samples))
+
+        def __len__(self):
+            return len(self.data)
+
+        def __getitem__(self, idx):
+            return self.data[idx]
+
+    dataloader = torch.utils.data.DataLoader(
+        dataset=DummyDataset(),
+        batch_sampler=nig,
+        num_workers=num_workers,
+    )
+    for _, expected, batch in zip(range(num_updates), nig_copy, dataloader):
+        assert np.array_equal(expected, batch.numpy())