+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- fix: `execute_experiment` accepts `None` as `evaluation_strategy` (#83)
+
 ### Added
 
 ## [0.7.0] - 2022-02-01