+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.8.0] - 2022-03-01
+
 ### Fixed
 
 - fix: `execute_experiment` has no side effects on its arguments (#90)